backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
websockets==12.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
smartapi-python
pyotp
logzero
websocket-client
pycryptodome